APIMF-3208: rename AMFGraphClient to AMFGraphDocumentClient for consistency
