+++ shared/src/main/scala/amf/core/client/platform/AMFGraphDocumentClient.scala
@@ -1,32 +1,35 @@
-package amf.core.client.scala
+package amf.core.client.platform
 
-import amf.core.client.scala.parse.AMFParser
-import amf.core.client.scala.render.AMFRenderer
-import amf.core.client.scala.transform.AMFTransformer
-import amf.core.client.scala.validation.{AMFValidationReport, AMFValidator}
-import amf.core.client.scala.model.document.BaseUnit
+import amf.core.internal.convert.CoreClientConverters._
+import amf.core.client.platform.model.document.BaseUnit
+import amf.core.client.scala.{AMFGraphDocumentClient => InternalAMFGraphDocumentClient}
+import amf.core.client.platform.validation.AMFValidationReport
 import amf.core.client.common.validation.ProfileName
-import amf.core.client.scala.parse.document.ParsedDocument
 import org.yaml.builder.DocBuilder
-
-import scala.concurrent.{ExecutionContext, Future}
+import scala.concurrent.ExecutionContext
+import scala.scalajs.js.annotation.{JSExportAll, JSExportTopLevel}
 
 /**
   * Contains common AMF graph operations.
   * Base client for <code>AMLClient</code> and <code>AMFClient</code>.
   */
-class AMFGraphClient(protected val configuration: AMFGraphConfiguration) {
+@JSExportAll
+class AMFGraphDocumentClient private[amf] (private val _internal: InternalAMFGraphDocumentClient) {
 
-  implicit val exec: ExecutionContext = configuration.getExecutionContext
+  private[amf] def this(configuration: AMFGraphConfiguration) = {
+    this(new InternalAMFGraphDocumentClient(configuration))
+  }
 
-  def getConfiguration: AMFGraphConfiguration = configuration
+  private implicit val ec: ExecutionContext = _internal.getConfiguration.getExecutionContext
+
+  def getConfiguration(): AMFGraphConfiguration = _internal.getConfiguration
 
   /**
     * Asynchronously generate a BaseUnit from the content located in the given url.
     * @param url Location of the file to parse
     * @return A CompletableFuture of [[AMFResult]]
     */
-  def parse(url: String): Future[AMFResult] = AMFParser.parse(url, configuration)
+  def parse(url: String): ClientFuture[AMFResult] = _internal.parse(url).asClient
 
   /**
     * Asynchronously generate a BaseUnit from the content located in the given url.
@@ -35,14 +38,14 @@ class AMFGraphClient(protected val configuration: AMFGraphConfiguration) {
     *                  Examples: <code>"application/raml10"</code> or <code>"application/raml10+yaml"</code>
     * @return A CompletableFuture of [[AMFResult]]
     */
-  def parse(url: String, mediaType: String): Future[AMFResult] = AMFParser.parse(url, mediaType, configuration)
+  def parse(url: String, mediaType: String): ClientFuture[AMFResult] = _internal.parse(url, mediaType).asClient
 
   /**
     * Asynchronously generate a BaseUnit from a given string.
     * @param content The content as a string
     * @return A CompletableFuture of [[AMFResult]]
     */
-  def parseContent(content: String): Future[AMFResult] = AMFParser.parseContent(content, configuration)
+  def parseContent(content: String): ClientFuture[AMFResult] = _internal.parseContent(content).asClient
 
   /**
     * Asynchronously generate a BaseUnit from a given string.
@@ -51,15 +54,15 @@ class AMFGraphClient(protected val configuration: AMFGraphConfiguration) {
     *                  Examples: <code>"application/raml10"</code> or <code>"application/raml10+yaml"</code>
     * @return A CompletableFuture of [[AMFResult]]
     */
-  def parseContent(content: String, mediaType: String): Future[AMFResult] =
-    AMFParser.parseContent(content, mediaType, configuration)
+  def parseContent(content: String, mediaType: String): ClientFuture[AMFResult] =
+    _internal.parseContent(content, mediaType).asClient
 
   /**
     * Transforms a [[BaseUnit]] with the default configuration
     * @param bu [[BaseUnit]] to transform
     * @return An [[AMFResult]] with the transformed BaseUnit and it's report
     */
-  def transform(bu: BaseUnit): AMFResult = AMFTransformer.transform(bu, configuration) // clone? BaseUnit.resolved
+  def transform(bu: BaseUnit): AMFResult = _internal.transform(bu)
 
   /**
     * Transforms a [[BaseUnit]] with a specific pipeline
@@ -67,22 +70,14 @@ class AMFGraphClient(protected val configuration: AMFGraphConfiguration) {
     * @param pipelineName name of any custom or [[AMFGraphConfiguration.predefined predefined]] pipeline
     * @return An [[AMFResult]] with the transformed BaseUnit and it's report
     */
-  def transform(bu: BaseUnit, pipelineName: String): AMFResult =
-    AMFTransformer.transform(bu, pipelineName, configuration) // clone? BaseUnit.resolved
+  def transform(bu: BaseUnit, pipelineName: String): AMFResult = _internal.transform(bu, pipelineName)
 
   /**
     * Render a [[BaseUnit]] to its default type
     * @param bu [[BaseUnit]] to be rendered
     * @return The content rendered
     */
-  def render(bu: BaseUnit): String = AMFRenderer.render(bu, configuration)
-
-  /**
-    * Render a [[BaseUnit]] and return the AST
-    * @param bu [[BaseUnit]] to be rendered
-    * @return the AST as a [[ParsedDocument]]
-    */
-  def renderAST(bu: BaseUnit): ParsedDocument = AMFRenderer.renderAST(bu, configuration)
+  def render(bu: BaseUnit): String = _internal.render(bu)
 
   /**
     * Render a [[BaseUnit]] to a certain mediaType
@@ -91,16 +86,7 @@ class AMFGraphClient(protected val configuration: AMFGraphConfiguration) {
     *                  Examples: <code>"application/raml10"</code> or <code>"application/raml10+yaml"</code>
     * @return The content rendered
     */
-  def render(bu: BaseUnit, mediaType: String): String = AMFRenderer.render(bu, mediaType, configuration)
-
-  /**
-    * Render a [[BaseUnit]] to a certain mediaType and return the AST
-    * @param bu [[BaseUnit]] to be rendered
-    * @param mediaType The nature and format of the given content. Must be <code>"application/spec"</code> or <code>"application/spec+syntax"</code>.
-    *                  Examples: <code>"application/raml10"</code> or <code>"application/raml10+yaml"</code>
-    * @return the AST as a [[ParsedDocument]]
-    */
-  def renderAST(bu: BaseUnit, mediaType: String): ParsedDocument = AMFRenderer.renderAST(bu, mediaType, configuration)
+  def render(bu: BaseUnit, mediaType: String): String = _internal.render(bu, mediaType)
 
   /**
     * Render a [[BaseUnit]] to a [[DocBuilder]] in the form of a graph (jsonld)
@@ -108,15 +94,14 @@ class AMFGraphClient(protected val configuration: AMFGraphConfiguration) {
     * @param builder [[DocBuilder]] which is used for rendering
     * @return The result produced by the DocBuilder after rendering
     */
-  def renderGraphToBuilder[T](bu: BaseUnit, builder: DocBuilder[T]): T =
-    AMFRenderer.renderGraphToBuilder(bu, builder, configuration)
+  def renderGraphToBuilder[T](bu: BaseUnit, builder: DocBuilder[T]): T = _internal.renderGraphToBuilder(bu, builder)
 
   /**
     * Validate a [[BaseUnit]] with its default validation profile name
     * @param bu [[BaseUnit]] to validate
     * @return an [[AMFValidationReport]]
     */
-  def validate(bu: BaseUnit): Future[AMFValidationReport] = AMFValidator.validate(bu, configuration)
+  def validate(bu: BaseUnit): ClientFuture[AMFValidationReport] = _internal.validate(bu).asClient
 
   /**
     * Validate a [[BaseUnit]] with a specific validation profile name
@@ -124,6 +109,6 @@ class AMFGraphClient(protected val configuration: AMFGraphConfiguration) {
     * @param profileName the [[ProfileName]] of the desired validation profile
     * @return an [[AMFValidationReport]]
     */
-  def validate(bu: BaseUnit, profileName: ProfileName): Future[AMFValidationReport] =
-    AMFValidator.validate(bu, profileName, configuration)
+  def validate(bu: BaseUnit, profileName: ProfileName): ClientFuture[AMFValidationReport] =
+    _internal.validate(bu, profileName).asClient
 }
