+++ shared/src/test/scala/amf/core/render/AmfGraphRendererTest.scala
@@ -30,7 +30,7 @@ trait AmfGraphRendererTest
           AMFGraphConfiguration
             .predefined()
             .withRenderOptions(RenderOptions().withPrettyPrint.withoutFlattenedJsonLd)
-            .createClient()
+            .documentClient()
             .render(document, AMF.mediaType))
       file   <- writeTemporaryFile(golden)(rendered)
       result <- assertDifferences(file, golden)
